setup/pyproject.toml
[project]
name = "setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.9",
    "python-dotenv>=1.1.1",
]
